Move learner's DLs to devices after parallelizing
This change takes advantage of and relies on the work done in the
`device` property setter of DataLoader. That setter moves the
DataLoader's Transformers over to the correct device.

For example, if this change were instead:

	self.learn.dls.loaders = [self._wrap_dl(to_device(dl, self.learn.model.device)) for dl in self.dls]

Then the DataLoader's Transformers might be on a different device than
the DataLoader that it belongs to. If the Transformer's `encodes` method
relies on state that was initialized before the Transformer was
distributed, then that state won't be on the same device of the batch
that it's encoding.